move worker credential storage test to worker_test.go to avoid import cycle
