+++ internal/daemon/worker/testing.go
@@ -174,7 +174,7 @@ type TestWorkerOpts struct {
 	WorkerAuthKms wrapping.Wrapper
 
 	// The worker credential storage KMS to use, or one will be created
-	WorkerStorage wrapping.Wrapper
+	WorkerStorageKms wrapping.Wrapper
 
 	// The name to use for the worker, otherwise one will be randomly
 	// generated, unless provided in a non-nil Config
@@ -263,6 +263,8 @@ func NewTestWorker(t testing.TB, opts *TestWorkerOpts) *TestWorker {
 	switch {
 	case opts.WorkerAuthKms != nil:
 		tw.b.WorkerAuthKms = opts.WorkerAuthKms
+	case opts.WorkerStorageKms != nil:
+		tw.b.WorkerStorageKms = opts.WorkerStorageKms
 	default:
 		if err := tw.b.SetupKMSes(tw.b.Context, nil, opts.Config); err != nil {
 			t.Fatal(err)
